requirements.txt
Markdown==3.3.3
Flask==1.1.1
Flask_SQLAlchemy==2.4.4
secrets==1.0.2
PyJWT==1.7.1
python_bcrypt==0.3.2
